src/requirements.txt
cloudshell-tg-trex>=1.0,<1.1
cloudshell-shell-core>=3.1,<3.2